Move external call handlers to a separate code file
Also:
1. Added MessageBoxOptions.ServiceNotification to all Message Boxes that shown inside
external call handlers. Without this option, message boxes are shown in the background.
2. Replace EnhancedContextMaker with SimpleContextMaker to speed up display of a form.
